pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "hk-stock-allocation"
version = "0.1.0"
description = "HK portfolio allocation and valuation signals using RQData"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "openpyxl>=3.1",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "rqdatac>=3.4",
]

[project.scripts]
hk-alloc = "hk_alloc.main:cli"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "integration: tests that require live RQData credentials and network access",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hk_alloc"]
